Refactor posnorm.sh.  Minimize transformations by setting up a target space in flipreg, combining all dofs within flipreg, and transforming only once in preparation for flipped registration.  The preliminary transformation is now built from the output of the centering (with -cog) or MNI-aligning.  Translations are extracted and combined with downscaling.  This ensures that no part of the image gets cut off from being moved outside the target grid.
Change interpolation on applying the preliminary transformation to
sinc.  This increases both accuracy and runtime.  It also makes
posnorm.sh (even) more specific to MR images.  At some point, the
interpolation mode will be made selectable.

Quantitative results are substantially better now.  Most likely this
is mainly due to the earlier bug fix that eliminated duplicate
application of the centering transform.

A future addition may be a testing mode.  For now, the recommended
testing procedure is to generate an image with perfect symmetry
(e.g. by flipping an image and summation of the original with the
flipped).  Then perturb with a random transformation.  Run posnorm.sh
on the output and compare the aligned image with the non-perturbed
image, and the recovered transformation with the inverse of the
perturbing transformation.